Refactgor schema into separate files
