+++ shared/schema/index.ts
@@ -1,0 +1,9 @@
+export * from './content';
+export * from './sessions';
+export * from './platforms';
+export * from './content-platforms';
+export * from './users';
+export * from './issues';
+export * from './watchlist';
+export * from './api-usage';
+export * from './subgenres';
