+++ shared/schema/sessions.ts
@@ -1,0 +1,11 @@
+import { pgTable, varchar, jsonb, timestamp, index } from 'drizzle-orm/pg-core';
+
+export const sessions = pgTable(
+  'sessions',
+  {
+    sid: varchar('sid').primaryKey(),
+    sess: jsonb('sess').notNull(),
+    expire: timestamp('expire').notNull(),
+  },
+  (table) => [index('IDX_session_expire').on(table.expire)]
+);
