+++ shared/schema/watchlist.ts
@@ -1,0 +1,29 @@
+import { pgTable, serial, integer, timestamp, index } from 'drizzle-orm/pg-core';
+import { createInsertSchema } from 'drizzle-zod';
+import { z } from 'zod';
+
+import { users } from './users';
+import { content } from './content';
+
+export const watchlist = pgTable(
+  'watchlist',
+  {
+    id: serial('id').primaryKey(),
+    userId: integer('user_id')
+      .notNull()
+      .references(() => users.id, { onDelete: 'cascade' }),
+    contentId: integer('content_id')
+      .notNull()
+      .references(() => content.id, { onDelete: 'cascade' }),
+    createdAt: timestamp('created_at').defaultNow().notNull(),
+  },
+  (table) => [index('IDX_user_content_unique').on(table.userId, table.contentId)]
+);
+
+export const insertWatchlistSchema = createInsertSchema(watchlist).omit({
+  id: true,
+  createdAt: true,
+});
+
+export type InsertWatchlist = z.infer<typeof insertWatchlistSchema>;
+export type Watchlist = typeof watchlist.$inferSelect;
